More moving of stuff to dedicated arepo frontend
