+++ yt/frontends/gadget/data_structures.py
@@ -651,40 +651,3 @@ class GadgetHDF5Dataset(GadgetDataset):
             pass
 
         return valid
-
-
-class ArepoHDF5Dataset(GadgetHDF5Dataset):
-    _field_info_class = ArepoFieldInfo
-
-    def __init__(self, filename, dataset_type="arepo_hdf5",
-                 unit_base=None,
-                 smoothing_factor=2.0,
-                 index_order=None,
-                 index_filename=None,
-                 kernel_name=None,
-                 bounding_box=None,
-                 units_override=None,
-                 unit_system="cgs"):
-        super(ArepoHDF5Dataset, self).__init__(
-            filename, dataset_type=dataset_type, unit_base=unit_base,
-            index_order=index_order, index_filename=index_filename,
-            kernel_name=kernel_name, bounding_box=bounding_box,
-            units_override=units_override, unit_system=unit_system)
-        self.smoothing_factor = smoothing_factor
-
-    @classmethod
-    def _is_valid(self, *args, **kwargs):
-        need_groups = ['Header', 'Config']
-        veto_groups = ['FOF', 'Group', 'Subhalo']
-        valid = True
-        try:
-            fh = h5py.File(args[0], mode='r')
-            valid = all(ng in fh["/"] for ng in need_groups) and \
-                    not any(vg in fh["/"] for vg in veto_groups) and \
-                    ("VORONOI" in fh["/Config"].attrs.keys() or
-                     "AMR" in fh["/Config"].attrs.keys())
-            fh.close()
-        except:
-            valid = False
-            pass
-        return valid
